Refactor pkg and cmd
