+++ go.mod
@@ -7,5 +7,7 @@ require (
 	github.com/go-playground/validator v9.31.0+incompatible
 	github.com/gorilla/mux v1.8.0
 	github.com/leodido/go-urn v1.2.1 // indirect
+	go.opentelemetry.io/otel/exporters/stdout v0.18.0
+	go.opentelemetry.io/otel/sdk v0.18.0
 	gopkg.in/go-playground/assert.v1 v1.2.1 // indirect
 )
